Test configuration of EclipseWorkspace projects
The test is a copy of the parts of `DefaultModelRepositoryTest` which
handle `EclipseProject`s. This is to make sure that the projects
returned by the composite have the same behavior as if they were fetched
from a normal ProjectConnection. 

Eventually, fetching individual projects from the
`SimpleModelRepository` will be removed. Then we can get rid of these
duplicated tests, too.