Reduced logging levels for non-severe problems
Not being able to close a stream is not an error. It does not stop us
from continuing our work, it just might lead to file handle leaking.
Changed to warning.

Not being able to read a cache file is also not an error. It can be
recovered by recomputing the cached value. Only if that fails should we
throw an error. Changed to warning.