Add a CACHE_ONLY strategy to PublishedGradleVersions
Buildship uses the PublishedGradleVersions on the UI thread and should
only use the fast path from the cache there. The long-running update
from the server should only be done in the background. This change
provides the infrastructure needed for that improvement.

I also added a reasonable timeout for the HttpConnection, so we don't
try to connect for several minutes.