Make connection reuse an option instead of the default.
There is currently no use case for connection caching. The idea is that
in the future, connections could push changes to the client (instead of
just being used for pulling). This would require connections to be kept
open.

The downside of cached connections is that we don't detect changes to
the Gradle wrapper version, which has already been reported by several
users.

As long as we don't have the push-feature, connection reuse should be an
option, not the default.