Fix closing of connections with PER_REQUEST strategy
The previous implementation closed an exisiting connection when a new
one was requested. This works most of the time, as `connection.close()`
guarantees to run all remaining actions before returning. But it has two
problems:

- it waits synchronously for the other task to finish, so we can't e.g.
fetch a model while waiting for a build to finish
- it is prone to race conditions (see
https://bugs.eclipse.org/bugs/show_bug.cgi?id=492118)

The new strategy closes connections explicitly once their work is done
and does not affect other connections in any way. This solves both above
mentioned problems.