Upgrade TAPI to enable dependency substitution
Dependency substitution is currently not fluently enabled in the TAPI,
but instead has to be enabled explicitly using the 'integrated' flag.
The problem with that is that it treats all builds in the composite
using the same Gradle version. So to make sure that this will work, we
first need to check whether indeed all of those projects are using the
same Gradle version.

This check will be inside the TAPI itself in future versions and the
'integrated' flag will be removed.