Fix performance regression in composite build
The hierarchy computation in EclipseProject and GradleProject was done
over and over again for each project, as we get them in a flattened list
from the composite build API. Fixed by introducing a cache, which works
well because the list contains parent projects before their children.
This means that each hierarchy is only converted once.