Remove unused Environment enum from codebase
The enum was originally intriduced to support different models for the
Eclipse environments if 1) Eclipse was used, 2) the BuildInvocations
was queried and 3) old target Gradle version was specified.

Since we don't use the BuildInvocations for quite a while now, I removed
the traces of this Environment enum. A direct consequence is that the
integration build will finish twice as fast.